Address issue reported by Alex Greenbaum where hex_aperture behaved differently depending on even or odd npix values.
The best option we could come up with was picking only pixels
entirely enclosed in the hexagon for inclusion in the aperture. Now,
even and odd npix values alike will produce the behavior where the
corners on the non-flat sides lack a pixel that would take them to the
edge of the array. (In other words, hex_aperture arrays have a zero
column or row between the corner and the edge of the array.)